WIP implement lock-based protocol

+++ args.py
@@ -6,7 +6,7 @@ def get_args():
     parser.add_argument("--nodes", type=int, default=100)
     parser.add_argument("--runs-per-process", type=int, default=1000)
     parser.add_argument("--time-per-run", type=float, default=7)
-    parser.add_argument("--poisson-rate", type=int, default=1.0)
+    parser.add_argument("--poisson-rate", type=float, default=1.0)
     parser.add_argument("--k", type=int, default=7)
     parser.add_argument("--seed", type=int, default=42)
     parser.add_argument("--cpus", type=int, default=None)
